+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Сhangelog
 
+- Added support for a new exchange: [Bleutrade](https://bleutrade.com) (Brazil)
 - Added `.hasPublicAPI / .hasPrivateAPI` properties to all derived exchanges
 - Added support for a new exchange: [ACX](https://acx.io) (Australia)
 - Fixed Bitflyer's private GET APIs, [#187](https://github.com/kroitor/ccxt/pull/187)